requirements.txt
psycopg2-binary==2.9.9
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.0.0
google-generativeai==0.7.0

